# remove libc usage