--- a/snmalloc-sys/Cargo.toml
+++ b/snmalloc-sys/Cargo.toml
@@ -16,11 +16,6 @@ build = "build.rs"
 cc = {version = "1.0",optional=true}
 cmake = {version = "0.1",optional=true}
 
-
-[dependencies.libc]
-version = "0.2"
-default-features = false
-
 [features]
 default = ["1mib","build_cmake"]
 build_cc = ["cc"]