Not sure I see a reason to using libc::size_t instead of rust's usize which should be the same on rust platforms anyway. Saves 6 seconds on compilation and seems to reduce test time by .06 seconds because I believe libc is just overhead.

**Original:**
(base) D:\fork\snmalloc-rs>cargo test --workspace
    Updating crates.io index
   Compiling cc v1.0.67
   Compiling libc v0.2.93
   Compiling cmake v0.1.45
   Compiling snmalloc-sys v0.2.26 (D:\fork\snmalloc-rs\snmalloc-sys)
   Compiling snmalloc-rs v0.2.26 (D:\fork\snmalloc-rs)
    Finished test [unoptimized + debuginfo] target(s) in 29.75s
     Running target\debug\deps\snmalloc_rs-2847cd7b054242c4.exe

running 3 tests
test tests::it_frees_allocated_memory ... ok
test tests::it_frees_reallocated_memory ... ok
test tests::it_frees_zero_allocated_memory ... ok

test result: ok. 3 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.00s

     Running target\debug\deps\snmalloc_sys-0a5a449d1bd6b08e.exe

running 5 tests
test tests::it_frees_memory_malloc ... ok
test tests::it_calculates_usable_size ... ok
test tests::it_frees_memory_sn_malloc ... ok
test tests::it_frees_memory_sn_realloc ... ok
test tests::it_reallocs_correctly ... ok

test result: ok. 5 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.00s

   Doc-tests snmalloc-sys

running 0 tests

test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.00s

   Doc-tests snmalloc-rs

running 1 test
test src\lib.rs - (line 24) ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.57s

**New without libc:**
(base) D:\fork\snmalloc-rs>cargo test --workspace
   Compiling cc v1.0.67
   Compiling cmake v0.1.45
   Compiling snmalloc-sys v0.2.26 (D:\fork\snmalloc-rs\snmalloc-sys)
   Compiling snmalloc-rs v0.2.26 (D:\fork\snmalloc-rs)
    Finished test [unoptimized + debuginfo] target(s) in 23.36s
     Running target\debug\deps\snmalloc_rs-cafe9d9e8c56b241.exe

running 3 tests
test tests::it_frees_allocated_memory ... ok
test tests::it_frees_zero_allocated_memory ... ok
test tests::it_frees_reallocated_memory ... ok

test result: ok. 3 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.00s

     Running target\debug\deps\snmalloc_sys-52e4284250df16ff.exe

running 5 tests
test tests::it_calculates_usable_size ... ok
test tests::it_frees_memory_malloc ... ok
test tests::it_frees_memory_sn_malloc ... ok
test tests::it_frees_memory_sn_realloc ... ok
test tests::it_reallocs_correctly ... ok

test result: ok. 5 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.00s

   Doc-tests snmalloc-sys

running 0 tests

test result: ok. 0 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.00s

   Doc-tests snmalloc-rs

running 1 test
test src\lib.rs - (line 24) ... ok

test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out; finished in 0.51s