@SchrodingerZhu 2021-04-22
the unsafe guidelines(https://rust-lang.github.io/unsafe-code-guidelines/layout/scalars.html) say that usize==uintptr_t. but the C standard doesn't define that uintptr_t==size_t therefore usize != size_t.

But seems that the assumption is true on almost all the platforms. and we are talking about memory and address？ cc @mjp41.