@SchrodingerZhu 2021-04-22
after discussing with others, the only special case they find is that on Intel 8086 the managed memory size (size_t = 16bits) is less than the remote address size (*ptr_t = 32 bits).

Considering the use case of snmalloc, I think we can ignore this case? Let us wait to see how upstream suggests.