@mjp41 2021-04-22
Well, on CHERI the `uintptr_t` is a capability pointer so is 128bits, but the maximum allocation size is 64bits.  So `size_t` is different from `uintptr_t`.  I expect the Rust spec for this has not taken into account CHERI.  In some sense, C is lucky to have has some strange early architectures as it meant `size_t` and `uintptr_t` aren't conflated in the standard.

As you might have seen on `snmalloc`, we (well @nwf) are building support for CHERI, so I would be concerned about this conflation.   However, I think the problem lies in the unsafe guidelines and we should be able to drop the libc dependence.  

@davidchisnall, @rwatson, @nwf: I think the  Rust unsafe guidelines should be CHERI aware. Anyone want to raise an issue on the topic?