@davidchisnall 2021-04-22
I had a conversation at ASPLOS about this a few years back.  Apparently someone had raised this with the Rust community when they were standardising `usize`, but they had ignored the concerns.  I think we're going to have some issues with Rust code coexisting with memory-safety technology, given the conflation of 'integer that can represent the size of any object' and 'integer type through which you can losslessly round-trip a pointer'.  I found it quite sad that a language that aims for safety would make this conflation.  It will also be a problem for Rust on some of the proposed WebAssembly extensions that allow incremental memory safety.

One of the companies competing for the next round of Digital Security by Design funding expressed an interest in doing a Rust port to Morello, so I'd hope that they'd address this.  

All of that said, these uses are all places where CHERI C would use `size_t`, not `[u]intptr_t` and so Rust's USize is probably the right thing.  They describe sizes and alignments[1] and so are not things that expect to carry pointer provenance.  Rust will need to figure out a way to represent a real `[u]intptr_t` - it definitely shouldn't be called `usize`.

[1] I personally think that it's a bad idea to design an API that expresses alignment as anything other than a log2 value, but that's a separate issue.