@jacobbramley 2021-04-22
I think this is the most recent issue regarding `usize` in Rust+CHERI: https://github.com/rust-lang/rust/issues/65473