@ryancinsight 2021-04-23
While I believe rust will need to update it's definition of usize and how it will handle Cheri, I don't believe libc fixes anything and just appears redundant to me.

According to libc crate:
Type definition libc::size_t
Type size_t = usize;